# docs/swagger.yaml
basePath: /api
definitions:
...
  models.CreateNoteCategoryRequest:
    properties:
      color:
        type: string
      home_id:
        type: integer
      icon:
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - home_id
    - name
    type: object
  models.CreateNoteRequest:
    properties:
      content:
        type: string
      mentioned_bill_category_ids:
        items:
          type: integer
        type: array
      mentioned_bill_ids:
        items:
          type: integer
        type: array
      mentioned_note_category_ids:
        items:
          type: integer
        type: array
      mentioned_shopping_category_ids:
        items:
          type: integer
        type: array
      mentioned_shopping_item_ids:
        items:
          type: integer
        type: array
      mentioned_task_ids:
        items:
          type: integer
        type: array
      mentioned_user_ids:
        items:
          type: integer
        type: array
      note_category_id:
        type: integer
      title:
        maxLength: 256
        minLength: 1
        type: string
    required:
    - title
    type: object
...
  models.UpdateNoteCategoryRequest:
    properties:
      color:
        type: string
      icon:
        type: string
      name:
        maxLength: 64
        minLength: 2
        type: string
    type: object
  models.UpdateNoteRequest:
    properties:
      content:
        type: string
      mentioned_bill_category_ids:
        items:
          type: integer
        type: array
      mentioned_bill_ids:
        items:
          type: integer
        type: array
      mentioned_note_category_ids:
        items:
          type: integer
        type: array
      mentioned_shopping_category_ids:
        items:
          type: integer
        type: array
      mentioned_shopping_item_ids:
        items:
          type: integer
        type: array
      mentioned_task_ids:
        items:
          type: integer
        type: array
      mentioned_user_ids:
        items:
          type: integer
        type: array
      note_category_id:
        type: integer
      title:
        maxLength: 256
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Diploma project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Diploma Server API
  version: "1.0"
paths:
...
  /homes/{home_id}/note_categories:
    get:
      description: Get all note categories in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all note categories
      tags:
      - note_category
    post:
      consumes:
      - application/json
      description: Create a new category for notes in a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Create Category Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateNoteCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create note category
      tags:
      - note_category
  /homes/{home_id}/note_categories/{category_id}:
    delete:
      description: Delete a note category by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete note category
      tags:
      - note_category
    put:
      consumes:
      - application/json
      description: Update a note category by ID
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: integer
      - description: Update Category Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNoteCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update note category
      tags:
      - note_category
  /homes/{home_id}/notes:
    get:
      description: Get all notes for a home, optionally filtered by note category
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Category ID filter
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all notes
      tags:
      - note
    post:
      consumes:
      - application/json
      description: Create a new note in a home, optionally mentioning users, tasks,
        bills, shopping items, and categories
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Create Note Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new note
      tags:
      - note
  /homes/{home_id}/notes/{note_id}:
    delete:
      description: Delete a note from a home
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a note
      tags:
      - note
    get:
      description: Get detailed information of a note including its mentions
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get note by ID
      tags:
      - note
    put:
      consumes:
      - application/json
      description: Update a note details and its mentions
      parameters:
      - description: Home ID
        in: path
        name: home_id
        required: true
        type: integer
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      - description: Update Note Request
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a note
      tags:
      - note
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"